refactor Header
